style: 반응형 레이아웃 조정
