+++ src/components/community/ReviewList.jsx
@@ -1,19 +1,11 @@
 import { useState } from "react";
 import { useQuery } from "react-query";
-
 import ReviewItem from "./ReviewItem";
-
 import Spinner from "components/common/Spinner";
 import Pagination from "components/common/pagination/Pagination";
-import CommunityButton from "./CommunityButton";
-import { MdEdit } from "react-icons/md";
-import { useNavigate } from "react-router-dom";
-import Swal from "sweetalert2";
-import SearchBox from "./SearchBox";
+import CommunityFooter from "./CommunityFooter";
 
 const ReviewList = ({ queryFn }) => {
-  const navigate = useNavigate();
-
   const postsPerPage = 10;
   const [page, setPage] = useState(1);
 
@@ -46,25 +38,10 @@ const ReviewList = ({ queryFn }) => {
     .fill()
     .map((_, index) => index + 1);
 
-  const goEdit = () => {
-    if (localStorage.getItem("accessToken") != null) {
-      navigate("/community/edit");
-    } else {
-      Swal.fire("로그인이 필요합니다!");
-    }
-  };
-
   return (
     <div className="m-4 mt-6">
       {content}
-
-      <div className="flex justify-between items-center flex-col md:flex-row">
-        <SearchBox />
-        <CommunityButton type="button" onClickHandler={goEdit}>
-          <MdEdit className="mr-1" />
-          쓰기
-        </CommunityButton>
-      </div>
+      <CommunityFooter />
       <Pagination
         page={page}
         setPage={setPage}
